# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faed Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success?: boolean;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  buyer: "buyer",
  supplier: "supplier",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: RegisterRequestRole;
  businessName?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  supplier: "supplier",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierStatus?: string | null;
}

export interface AuthResponse {
  user: User;
  token?: string;
}

export type SupplierVerificationStatus =
  (typeof SupplierVerificationStatus)[keyof typeof SupplierVerificationStatus];

export const SupplierVerificationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Supplier {
  id: number;
  userId: number;
  businessName: string;
  verificationStatus: SupplierVerificationStatus;
  userName?: string;
  userEmail?: string;
}

export interface Product {
  id: number;
  supplierId?: number;
  supplierName?: string;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  price: number;
  originalPrice: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  discountReason?: string | null;
  sizes?: string[];
  /** @nullable */
  brand?: string | null;
  isActive?: boolean;
  createdAt?: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
}

export interface CartItem {
  productId: number;
  quantity: number;
  product: Product;
}

export interface CartResponse {
  items: CartItem[];
  total: number;
}

export interface AddToCartRequest {
  productId: number;
  quantity: number;
}

export interface UpdateCartRequest {
  quantity: number;
}

export type CreateOrderRequestPaymentMethod =
  (typeof CreateOrderRequestPaymentMethod)[keyof typeof CreateOrderRequestPaymentMethod];

export const CreateOrderRequestPaymentMethod = {
  mada: "mada",
  stcpay: "stcpay",
  tabby: "tabby",
  cod: "cod",
} as const;

export interface CreateOrderRequest {
  name: string;
  phone: string;
  address: string;
  paymentMethod: CreateOrderRequestPaymentMethod;
  /** @nullable */
  notes?: string | null;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName?: string;
  quantity: number;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  customerName?: string;
  status: OrderStatus;
  totalPrice: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  createdAt: string;
  items?: OrderItem[];
}

export interface ProductInput {
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  originalPrice: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  discountReason?: string | null;
  sizes?: string[];
  /** @nullable */
  brand?: string | null;
}

export interface SupplierStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
}

export interface AdminStats {
  totalUsers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingSuppliers: number;
}

export type ListProductsParams = {
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  size?: string;
  brand?: string;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};