# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Faed Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: My orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/suppliers/{id}/approve:
    patch:
      operationId: approveSupplier
      tags: [admin]
      summary: Approve supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
...
  /admin/products/{id}:
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Remove product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [buyer, supplier]
        businessName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [buyer, supplier, admin]
        supplierId:
          type: integer
          nullable: true
        supplierStatus:
          type: string
          nullable: true
    Supplier:
      type: object
      required: [id, userId, businessName, verificationStatus]
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessName:
          type: string
        verificationStatus:
          type: string
          enum: [pending, approved, rejected]
        userName:
          type: string
        userEmail:
          type: string
    Product:
      type: object
      required: [id, name, price, originalPrice, quantity, category]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string
          nullable: true
        discountReason:
          type: string
          nullable: true
        sizes:
          type: array
          items:
            type: string
        brand:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
...
    CartItem:
      type: object
      required: [productId, quantity, product]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
    CartResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
    AddToCartRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    UpdateCartRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    CreateOrderRequest:
      type: object
      required: [name, phone, address, paymentMethod]
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        paymentMethod:
          type: string
          enum: [mada, stcpay, tabby, cod]
        notes:
          type: string
          nullable: true
    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
    Order:
      type: object
      required: [id, userId, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        totalPrice:
          type: number
        paymentMethod:
          type: string
          nullable: true
        shippingAddress:
          type: string
          nullable: true
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...