# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faed Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestPaymentMethod } from "./createOrderRequestPaymentMethod";

export interface CreateOrderRequest {
  name: string;
  phone: string;
  address: string;
  paymentMethod: CreateOrderRequestPaymentMethod;
  /** @nullable */
  notes?: string | null;
}