# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faed Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  userId: number;
  customerName?: string;
  status: OrderStatus;
  totalPrice: number;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  createdAt: string;
  items?: OrderItem[];
}