# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faed Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  supplierId?: number;
  supplierName?: string;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  price: number;
  originalPrice: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  discountReason?: string | null;
  sizes?: string[];
  /** @nullable */
  brand?: string | null;
  isActive?: boolean;
  createdAt?: string;
}